client/requirements.txt
numpy==1.26.4
rpi_ws281x
adafruit-circuitpython-neopixel
RPi.GPIO
setuptools
wheel
spacy==3.0.0
blis==0.7.4
nltk==3.9b1
SpeechRecognition
pyaudio
requests